Create a laodTable.js file
